requirements.txt
aiogram 
openai==0.28
python-dotenv
